OAuth: Extract client authentication service to interface
